Replace *First convenience methods with functions
